requirements.txt
blosc==1.10.6
deep_pipeline==0.2.5
Flask==2.2.2
gevent==21.12.0
imgaug==0.4.0
matplotlib==3.4.2
monai==0.8.1
numpy==1.21.1
opencv_python_headless==4.6.0.66
pandas==1.3.2
progressbar33==2.4
pydicom==2.3.0
sas7bdat==2.2.3
scikit_image==0.16.2
scikit_learn==1.0.2
scipy==1.7.3
seaborn==0.11.2
SimpleITK==2.2.1
solt==0.1.8
torch==1.10.1
torchvision==0.11.2
